requirements.txt
crewai==0.28.8
langchain==0.1.20
langchain-openai==0.1.7
langchain-community==0.0.38
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
colorama==0.4.6
rich==13.7.1
pydantic==2.7.1
openai==1.30.1